requirements.txt
asgiref==3.11.1
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
pdfminer.six==20260107
pycparser==3.0
sqlparse==0.5.5
typing_extensions==4.15.0
